added seeder stuff
