+++ express-project-starter/db/seeders/20210504141134-answer-seed.js
@@ -1,29 +1,4 @@
 "use strict";
-const fetch = require("node-fetch");
-
-const theFetch = async () => {
-  return answer;
-};
-
-const seed = async () => {
-  let answers = [];
-  for (let i = 0; i < 10; i++) {
-    let ownerId = Math.floor(Math.random() * 10);
-    let questionId = Math.floor(Math.random() * 10);
-    const x = await fetch(
-      "https://baconipsum.com/api/?type=all-meat&paras=2&start-with-lorem=1"
-    );
-    let answer = await x.json();
-    answer = JSON.stringify(answer);
-    answer = await JSON.parse(answer);
-    // console.log(answer);
-    const obj = { answer: answer[0], ownerId, questionId, content: "" };
-    answers.push(obj);
-  }
-  return answers;
-};
-
-console.log(seed());
 
 module.exports = {
   up: (queryInterface, Sequelize) => {
@@ -32,11 +7,8 @@ module.exports = {
       Return a promise to correctly handle asynchronicity.
 
       Example:
-      return queryInterface.bulkInsert('People', [{
-        name: 'John Doe',
-        isBetaMember: false
-      }], {});
-    */
+      */
+    return queryInterface.bulkInsert("Answers", [], {});
   },
 
   down: (queryInterface, Sequelize) => {
@@ -45,7 +17,7 @@ module.exports = {
       Return a promise to correctly handle asynchronicity.
 
       Example:
-      return queryInterface.bulkDelete('People', null, {});
-    */
+      */
+    return queryInterface.bulkDelete("Answers", null, {});
   },
 };
